+++ application/models/game_model.php
@@ -274,6 +274,26 @@ class Game_model extends MY_Model
         return $result ? $result[0] : null;
     }
 
+    public function clearAllStageToPlayer($client_id, $site_id, $game_id, $pb_player_id)
+    {
+        $this->set_site_mongodb($site_id);
+
+        $this->mongo_db->where(array(
+            'client_id' => new MongoId($client_id),
+            'site_id' => new MongoId($site_id),
+            'game_id' => new MongoId($game_id),
+            'pb_player_id' => new MongoId($pb_player_id),
+            'is_current'=> true
+        ));
+
+        $d = new MongoDate(time());
+
+        $this->mongo_db->set('is_current', false );
+        $this->mongo_db->set('date_modified', $d);
+        $this->mongo_db->update('playbasis_game_stage_to_player');
+
+    }
+
     public function setStageToPlayer($client_id, $site_id, $game_id, $stage_level, $pb_player_id, $data)
     {
         $d = new MongoDate(time());
@@ -284,8 +304,6 @@ class Game_model extends MY_Model
         $insert_data['game_id'] = new MongoId($game_id);
         $insert_data['stage_level'] = (int)$stage_level;
         $insert_data['pb_player_id'] = new MongoId($pb_player_id);
-
-        $insert_data['harvested_item'] = (isset($data['harvested_item']) && $data['harvested_item']) ? $data['harvested_item'] : array();
 
         if(isset($data['is_current']) && $data['is_current']){
             $insert_data['is_current'] = (bool)$data['is_current'] ;
@@ -306,9 +324,6 @@ class Game_model extends MY_Model
             'stage_level' => (int)$stage_level,
             'pb_player_id' => new MongoId($pb_player_id)
         ));
-        if(isset($data['harvested_item']) && $data['harvested_item']){
-            $this->mongo_db->set('harvested_item', $data['harvested_item'] );
-        }
 
         if(isset($data['is_current'])){
             $this->mongo_db->set('is_current', (bool)$data['is_current'] );